Create `config::kafka` Module
